# api/openapi.yaml
openapi: 3.1.0
info:
  title: StackPilot API
  version: 0.0.0
  description: StackPilot local control-plane API contract.
servers:
  - url: /
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /api/v1/services/{systemId}/{serviceId}/logs:
    parameters:
      - $ref: "#/components/parameters/SystemId"
      - $ref: "#/components/parameters/ServiceId"
    get:
      operationId: getServiceLogs
      summary: Read a bounded newest-first historical log window
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdQuery"
        - name: instanceId
          in: query
          required: true
          schema:
            type: string
            pattern: "^si_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        - name: cursor
          in: query
          required: false
          description: Return the latest entries with sequence values lower than this cursor.
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 500
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [trace, debug, info, warn, error, fatal]
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Ascending log entries from the selected newest window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/log-stream:
    get:
      operationId: streamServiceLogs
      summary: Stream redacted service logs after a sequence cursor
      parameters:
        - name: instanceId
          in: query
          required: true
          schema:
            type: string
            pattern: "^si_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        - name: serviceId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ManifestId"
        - name: afterSequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A server-sent log stream whose SSE IDs are service-instance sequence values.
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Long-lived local token read only from OS-protected storage by CLI clients.
    sessionAuth:
      type: apiKey
      in: cookie
      name: stackpilot_session
  parameters:
...
    WorkspaceIdQuery:
      name: workspaceId
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/WorkspaceId"
    SystemId:
      name: systemId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ManifestId"
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ManifestId"
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - AUTH_BOOTSTRAP_INVALID
        - EVENT_CURSOR_EXPIRED
        - FEATURE_NOT_ENABLED
        - AUTH_BROWSER_REQUEST_REJECTED
        - AUTH_CAPACITY_REACHED
        - AUTH_SESSION_INVALID
        - AUTH_TOKEN_INVALID
        - CONTROL_PLANE_RESTARTED
        - COMPOSE_CONFIG_INVALID
        - COMPOSE_BUILD_CONFIG_INVALID
        - COMPOSE_BUILD_FAILED
        - COMPOSE_BUILD_TIMEOUT
        - COMPOSE_DISCOVERY_FAILED
        - COMPOSE_INSPECT_FAILED
        - COMPOSE_LOG_FOLLOW_FAILED
        - COMPOSE_LIFECYCLE_INVALID
        - COMPOSE_LIFECYCLE_TIMEOUT
        - COMPOSE_NOT_FOUND
        - COMPOSE_OVERRIDE_CONFLICT
        - COMPOSE_OVERRIDE_INVALID
        - COMPOSE_PREFLIGHT_TIMEOUT
        - COMPOSE_PROJECT_IDENTITY_MISMATCH
        - COMPOSE_PROJECT_NOT_FOUND
        - COMPOSE_START_FAILED
        - COMPOSE_STOP_FAILED
        - CONTAINER_UNHEALTHY
        - COMPOSE_VERSION_UNSUPPORTED
        - DOCKER_DAEMON_UNAVAILABLE
        - DOCKER_NOT_FOUND
        - DOCKER_VERSION_UNSUPPORTED
        - HEALTH_NOT_READY
        - HEALTH_READINESS_TIMEOUT
        - HTTP_BODY_MISMATCH
        - HTTP_STATUS_MISMATCH
        - HTTP_TIMEOUT
        - INTERNAL_ERROR
        - LOG_CURSOR_EXPIRED
        - IDEMPOTENCY_KEY_REUSED
        - MANIFEST_DUPLICATE_KEY
        - MANIFEST_CYCLE_DETECTED
        - MANIFEST_DURATION_INVALID
        - MANIFEST_FILE_TOO_LARGE
        - MANIFEST_HEALTH_TARGET_UNSAFE
        - MANIFEST_CHANGED
        - MANIFEST_MULTIPLE_DOCUMENTS
        - MANIFEST_NOT_REGULAR_FILE
        - MANIFEST_PATH_OUTSIDE_WORKSPACE
        - MANIFEST_PORT_RANGE_INVALID
        - MANIFEST_REFERENCE_NOT_FOUND
        - MANIFEST_SCHEMA_INVALID
        - MANIFEST_SEMANTIC_INVALID
        - MANIFEST_TEMPLATE_INVALID
        - MANIFEST_UNKNOWN_FIELD
        - MANIFEST_YAML_INVALID
        - METHOD_NOT_ALLOWED
        - OPERATION_ALREADY_ACTIVE
        - OPERATION_INVALID_STATE
        - OPERATION_NOT_CANCELLABLE
        - OPERATION_NOT_FOUND
        - PLATFORM_NOT_SUPPORTED
        - PORT_CONFLICT
        - PROCESS_EXITED
        - PROCESS_IDENTITY_MISMATCH
        - PROCESS_START_FAILED
        - PROCESS_STOP_FAILED
        - REQUEST_VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - RUNNER_NOT_FOUND
        - RUNNER_PATH_UNSAFE
        - RUNNER_VERSION_CHECK_FAILED
        - RUNNER_VERSION_CHECK_TIMEOUT
        - SECRET_INVALID
        - SECRET_NOT_FOUND
        - SECRET_VERSION_CONFLICT
        - TCP_REFUSED
        - TCP_TIMEOUT
        - SUPERVISOR_EXITED
        - SUPERVISOR_UNAVAILABLE
        - WORKSPACE_ALREADY_REGISTERED
        - WORKSPACE_ID_REQUIRED
        - WORKSPACE_MANIFEST_UNAVAILABLE
        - WORKSPACE_NOT_FOUND
        - WORKSPACE_PATH_INVALID
        - WORKSPACE_SCRIPT_NOT_FOUND
        - WORKSPACE_SCRIPT_OUTSIDE
        - WORKSPACE_SCRIPT_TYPE_UNSUPPORTED
        - WORKSPACE_SCRIPT_ENCODING_UNSUPPORTED
        - WORKSPACE_SCRIPT_TOO_LARGE
        - WORKSPACE_SCRIPT_SYNTAX_UNSUPPORTED
        - WORKSPACE_SCRIPT_DANGEROUS
        - WORKSPACE_SCRIPT_REFERENCE_CYCLE
        - WORKSPACE_IMPORT_INCOMPLETE
        - WORKSPACE_IMPORT_PORT_UNCONFIRMED
        - WORKSPACE_IMPORT_DEPENDENCY_UNCONFIRMED
        - WORKSPACE_IMPORT_SOURCE_CHANGED
        - WORKSPACE_IMPORT_DRAFT_EXPIRED
        - WORKSPACE_MANIFEST_CONFLICT
        - WORKSPACE_MANIFEST_WRITE_FAILED
        - WORKSPACE_EDIT_RUNTIME_ACTIVE
        - WORKSPACE_UNREGISTER_RUNTIME_ACTIVE
        - WORKSPACE_RELINK_SYSTEM_MISMATCH
        - WORKSPACE_SYSTEM_ID_CHANGED
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
        - traceId
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        operationId:
          type:
            - string
            - "null"
          pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        traceId:
          type: string
          pattern: "^tr_[0-9a-f]{32}$"
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    LogEntry:
      type: object
      additionalProperties: false
      required: [timestamp, systemId, instanceId, serviceId, stream, level, message, sequence, truncated]
      properties:
        timestamp:
          type: string
          format: date-time
        systemId:
          $ref: "#/components/schemas/ManifestId"
        instanceId:
          type: string
          pattern: "^si_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        serviceId:
          $ref: "#/components/schemas/ManifestId"
        stream:
          type: string
          enum: [stdout, stderr]
        level:
          type: string
        message:
          type: string
        operationId:
          type: string
          pattern: "^op_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
        sequence:
          type: integer
          minimum: 1
        truncated:
          type: boolean
    LogPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          maxItems: 5000
          items:
            $ref: "#/components/schemas/LogEntry"
        nextCursor:
          type: [integer, "null"]
          minimum: 1
...
    ManifestId:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
...
    WorkspaceId:
      type: string
      pattern: "^ws_[0-7][0-9A-HJKMNP-TV-Z]{25}$"
...
  responses:
...
    BadRequest:
      description: The request or manifest syntax is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: The request conflicts with current catalog state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...